Introduce :search command
Introduce the :search command, allowing user to effectively hook into
the same code paths as used by '/', which does not seem possible
otherwise.
We want this functionality for something as simple as searching for the
currently selected file, which seems to be utterly impossible without
such a dance. An intuitive way would be to define a user command such
as:
  > command search /^%a$

and invoke that, alas, search behavior there seems to be special cased
in that it does not add the search pattern to the search history. Good.
Lord.